Begin extracting into an actual structure
Too tired to finish this, but wanted to commit it.
ArgumentError still needs something to format it.
Then it needs to delegate to the lib code.
Then NoMethodError needs to use the lib code, too.

I'm not totally sold on the structure. It might be too much for now,
I don't really have the feedback to say its right.
But at least it will be better than what we have at present,
and can be iterated on and unit tested.