Add an example of what might be a nice formatting language
I don't want to render colours directly into the output string,
because they become impossible to work with (ansi escape sequences
lack abstractions that prevent you from being able to correctly
format and colour unless you maintain a virtual terminal screen).

I'd also like to be able to semantically specify the information,
and then separate the presentation, like CSS does for HTML. This way
the error could be rendered to the terminal, or HTML, or really
to anything that presents info, so long as we make an adapter for it.

Not sure that this is what I want, but its probably on the right track.

Note that format nodes are immutable, and therefore only ever added onto.
This has some implications on the design of the code (eg if you pass
it to a method, the method has to return it in order for those changes
to take effect).