Move unique pieces of formatting into polymorphic classes
Going to try to get them joined into the parsing objects, too,
I'm pretty sure that's where they go at this point.
Just need some way to talk abstractly about the info they have,
so that the formatting code doesn't have to be coupled
to the logic of the heuristic.