Implement NameError via NoMethodError
They're similar enough, at present,
that we can let the NoMethodError handle the NameError use case.
Once we actually start adding in real heuristics,
this probably won't suffice, because knowing that NoMethodError
was an actual method call (vs potentially a local var),
could be useful for the heuristic.

Closes #16
Part of Iteration 2 (#31)