Don't freeze the global heuristic array
I think it should be frozen, but there is not currently a way to
modify the list of heuristics, so while this is wrong, it at least
doesn't make it a total pita (ie opening Config up and editing its ivars)

If anyone finds themselves doing this, let me know so that I can
build an API for whatever you're trying to do (ie, do we just need to
stick them at the beginning of the array? At the end won't work, b/c
Exception would swallow them... maybe they need to go somewhere in
the middle? How are you identifying where to place them? Are you
removing heuristics, too? etc)